Add files via upload
This is a Machine Learning script that I constructed using Tensorflow Object Detection API. I successfully built a model that can detect 3 objects: my face, a phone and a pillow. The number of objects is limited because I wanted to build a prototype first. More generalized model might be finished in the future.

Also note, due to the extremely large folder size, I wasn't able to upload every dependencies to GitHub but only the Python Script. All the other folders remain on my OneDrive.